test : 타입변경에 따른 테스트 수정

+++ backend/team6/src/test/java/programmers/team6/domain/vacation/repository/JdbcVacationEligibilitiesRepositoryTest.java
@@ -132,8 +132,8 @@ class JdbcVacationEligibilitiesRepositoryTest {
 	}
 
 	private record TestVacationInfo(
-		int totalCount,
-		int useCount,
+		double totalCount,
+		double useCount,
 		int version,
 		int memberId
 	) {
